c++, coroutines: Instrument missing return_void UB.
[dcl.fct.def.coroutine] / 6 Note 1:
"If return_void is found, flowing off the end of a coroutine is equivalent
to a co_return with no operand. Otherwise, flowing off the end of a
coroutine results in undefined behavior."

Here we implement this as a check for sanitized returns and call the ubsan
instrumentation; if that is not enabled we mark this as unreachable (which
might trap depending on the target settings).

gcc/cp/ChangeLog:

	* coroutines.cc
	(cp_coroutine_transform::wrap_original_function_body): Instrument
	the case where control flows off the end of a coroutine and the
	user promise has no return_void entry.

Signed-off-by: Iain Sandoe <iain@sandoe.co.uk>